Move sig_trap to machine independent code
