+++ sys/gen/Makefile
@@ -5,7 +5,6 @@ TOPDIR = $(realpath ../..)
 SOURCES = \
 	boot.c \
 	pmap.c \
-	signal.c \
 	syscall.c \
 	thread.c
