+++ sys/kern/signal.c
@@ -582,3 +582,17 @@ int do_sigreturn(ucontext_t *ucp) {
 
   return EJUSTRETURN;
 }
+
+void sig_trap(ctx_t *ctx, signo_t sig, int code, void *addr, int trapno) {
+  proc_t *proc = proc_self();
+
+  ksiginfo_t ksi;
+  ksi.ksi_flags = KSI_TRAP;
+  ksi.ksi_signo = sig;
+  ksi.ksi_code = code;
+  ksi.ksi_addr = addr;
+  ksi.ksi_trap = trapno;
+
+  WITH_MTX_LOCK (&proc->p_lock)
+    sig_kill(proc, &ksi);
+}
